move shared kubernetesYAMLToResources into resource pkg as a common func
